automation/requirements.txt
google-generative-ai
feedparser
requests
beautifulsoup4
schedule
python-dotenv
sanity
praw
linkedin-api
